15_10

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: pgomez-r <pgomez-r@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/05 14:45:53 by pgomez-r          #+#    #+#             */
-/*   Updated: 2022/10/14 10:00:24 by pgomez-r         ###   ########.fr       */
+/*   Updated: 2022/10/14 16:10:19 by pgomez-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 void	*ft_memmove(void *dest, const void *src, size_t count)
 {
-	size_t	i;
+	int	i;
 
 	if (!dest && !src)
 		return (0);
 	if (dest < src)
 	{
 		i = 0;
-		while (i < count)
+		while (i < (int)count)
 		{
 			((unsigned char *)dest)[i] = ((unsigned char *)src)[i];
 			i++;
